Start renaming to "item" instead of "package"
This switches to using the generic term "item" instead of "package",
and updates the items that appear in repositories to be members of
a sealed Item type.

+++ au.org.libraryforall.updater.inventory.vanilla/src/main/java/au/org/libraryforall/updater/inventory/vanilla/NamedVersion.kt
@@ -2,4 +2,4 @@ package au.org.libraryforall.updater.inventory.vanilla
 
 internal data class NamedVersion(
   val name: String,
-  val code: Int)
+  val code: Long)
